rename pdfviewer file
